IOS-5020 Renamed favorites to pin
