--- Anytype/Sources/PresentationLayer/Modules/HomeWidgets/Widgets/List/Specific/Favorite/FavoriteCompactListWidgetSubmoduleView.swift
@@ -1,11 +1,0 @@
-import Foundation
-import SwiftUI
-
-struct FavoriteCompactListWidgetSubmoduleView: View {
-    
-    let data: WidgetSubmoduleData
-    
-    var body: some View {
-        FavoriteCommonListWidgetSubmoduleView(data: data, style: .compactList)
-    }
-}
