--- Anytype/Sources/PresentationLayer/Modules/HomeWidgets/Widgets/List/Specific/Favorite/FavoriteListWidgetSubmoduleView.swift
@@ -1,11 +1,0 @@
-import Foundation
-import SwiftUI
-
-struct FavoriteListWidgetSubmoduleView: View {
-    
-    let data: WidgetSubmoduleData
-    
-    var body: some View {
-        FavoriteCommonListWidgetSubmoduleView(data: data, style: .list)
-    }
-}
