+++ Anytype/Sources/PresentationLayer/Modules/HomeWidgets/Widgets/List/Specific/Pinned/PinnedCommonListWidgetSubmoduleView.swift
@@ -1,0 +1,36 @@
+import Foundation
+import SwiftUI
+
+struct PinnedCommonListWidgetSubmoduleView: View {
+    
+    let data: WidgetSubmoduleData
+    let style: ListWidgetStyle
+    
+    var body: some View {
+        PinnedCommonListWidgetSubmoduleInternalView(data: data, style: style)
+            .id(data.widgetBlockId + style.rawValue)
+    }
+}
+
+private struct PinnedCommonListWidgetSubmoduleInternalView: View {
+    
+    let data: WidgetSubmoduleData
+    let style: ListWidgetStyle
+    
+    @StateObject private var model: PinnedWidgetInternalViewModel
+    
+    init(data: WidgetSubmoduleData, style: ListWidgetStyle) {
+        self.data = data
+        self.style = style
+        self._model = StateObject(wrappedValue: PinnedWidgetInternalViewModel(data: data))
+    }
+    
+    var body: some View {
+        ListWidgetView(
+            data: data,
+            style: style,
+            internalModel: model,
+            internalHeaderModel: nil
+        )
+    }
+}
