+++ Anytype/Sources/PresentationLayer/Modules/HomeWidgets/Widgets/List/Specific/Pinned/PinnedCompactListWidgetSubmoduleView.swift
@@ -1,0 +1,11 @@
+import Foundation
+import SwiftUI
+
+struct PinnedCompactListWidgetSubmoduleView: View {
+    
+    let data: WidgetSubmoduleData
+    
+    var body: some View {
+        PinnedCommonListWidgetSubmoduleView(data: data, style: .compactList)
+    }
+}
