+++ Anytype/Sources/PresentationLayer/Modules/HomeWidgets/Widgets/List/Specific/Pinned/PinnedListWidgetSubmoduleView.swift
@@ -1,0 +1,11 @@
+import Foundation
+import SwiftUI
+
+struct PinnedListWidgetSubmoduleView: View {
+    
+    let data: WidgetSubmoduleData
+    
+    var body: some View {
+        PinnedCommonListWidgetSubmoduleView(data: data, style: .list)
+    }
+}
