+++ Anytype/Sources/PresentationLayer/Modules/WidgetObjectList/Common/Buildes/WidgetObjectListMenuBuilder.swift
@@ -76,15 +76,15 @@ final class WidgetObjectListMenuBuilder: WidgetObjectListMenuBuilderProtocol {
         
         return .builder {
             
-            if allowOptions.contains(.favorite), isUndavoriteIds.isNotEmpty {
-                Action(optionTitle: Loc.favorite, optionImage: .X32.Favorite.favorite, menuTitle: Loc.addToFavorite, negative: false, action: { [weak output] in
-                    output?.setFavorite(objectIds: isUndavoriteIds, true)
+            if allowOptions.contains(.pin), isUndavoriteIds.isNotEmpty {
+                Action(optionTitle: Loc.pin, optionImage: .X32.Favorite.favorite, menuTitle: Loc.addToFavorite, negative: false, action: { [weak output] in
+                    output?.setPin(objectIds: isUndavoriteIds, true)
                 })
             }
             
-            if allowOptions.contains(.unfavorite), isFavoriteIds.isNotEmpty {
-                Action(optionTitle: Loc.unfavorite, optionImage: .X32.Favorite.unfavorite, menuTitle: Loc.removeFromFavorite, negative: false, action: { [weak output] in
-                    output?.setFavorite(objectIds: isFavoriteIds, false)
+            if allowOptions.contains(.unpin), isFavoriteIds.isNotEmpty {
+                Action(optionTitle: Loc.unpin, optionImage: .X32.Favorite.unfavorite, menuTitle: Loc.removeFromFavorite, negative: false, action: { [weak output] in
+                    output?.setPin(objectIds: isFavoriteIds, false)
                 })
             }
