+++ Anytype/Sources/PresentationLayer/Modules/WidgetObjectList/Common/Buildes/WidgetObjectListMenuOutput.swift
@@ -3,7 +3,7 @@ import Services
 
 @MainActor
 protocol WidgetObjectListMenuOutput: AnyObject, Sendable {
-    func setFavorite(objectIds: [String], _ isFavorite: Bool)
+    func setPin(objectIds: [String], _ isPinned: Bool)
     func setArchive(objectIds: [String], _ isArchived: Bool)
     func delete(objectIds: [String])
     func forceDelete(objectIds: [String])
