+++ Anytype/Sources/PresentationLayer/Modules/WidgetObjectList/Common/Internal/WidgetObjectListInternalViewModelProtocol.swift
@@ -35,7 +35,7 @@ protocol WidgetObjectListInternalViewModelProtocol: AnyObject {
 }
 
 extension WidgetObjectListInternalViewModelProtocol {
-    var availableMenuItems: [WidgetObjectListMenuItem] { [.favorite, .unfavorite, .moveToBin] }
+    var availableMenuItems: [WidgetObjectListMenuItem] { [.pin, .unpin, .moveToBin] }
     func onMove(from: IndexSet, to: Int) {}
     func subtitle(for details: ObjectDetails) -> String? { return nil }
 }
