+++ Anytype/Sources/PresentationLayer/Modules/WidgetObjectList/Common/Models/WidgetObjectListMenuItem.swift
@@ -1,8 +1,8 @@
 import Foundation
 
 enum WidgetObjectListMenuItem: String {
-    case favorite
-    case unfavorite
+    case pin
+    case unpin
     case moveToBin
     case delete
     case restore
