+++ Anytype/Sources/PresentationLayer/Modules/WidgetObjectList/Files/WidgetObjectListFilesManagerViewModel.swift
@@ -19,7 +19,7 @@ final class WidgetObjectListFilesManagerViewModel: WidgetObjectListInternalViewM
         title: Loc.EmptyView.Default.title,
         subtitle: Loc.EmptyView.Default.subtitle
     )
-    let editorScreenData: EditorScreenData = .favorites(homeObjectId: "", spaceId: "") // Is not used
+    let editorScreenData: EditorScreenData = .pinned(homeObjectId: "", spaceId: "") // Is not used
     var rowDetailsPublisher: AnyPublisher<[WidgetObjectListDetailsData], Never> { $rowDetails.eraseToAnyPublisher()}
     let editMode: WidgetObjectListEditMode = .editOnly
     let availableMenuItems: [WidgetObjectListMenuItem] = [.forceDelete]
