+++ Modules/Assets/Sources/Assets/Resources/Assets.xcassets/DesignSystem/SystemWidgets/Favorites.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "Favorites.pdf",
+      "filename" : "System Widgets.svg",
       "idiom" : "universal"
     }
   ],
