--- Modules/Assets/Sources/Assets/Resources/Assets.xcassets/DesignSystem/x24/Favorite/Contents.json
@@ -1,9 +1,0 @@
-{
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "properties" : {
-    "provides-namespace" : true
-  }
-}
