--- Modules/Assets/Sources/Assets/Resources/Assets.xcassets/DesignSystem/x24/Favorite/Favorite.imageset/Contents.json
@@ -1,15 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "Favorite.pdf",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "properties" : {
-    "template-rendering-intent" : "template"
-  }
-}
