--- Modules/Assets/Sources/Assets/Resources/Assets.xcassets/DesignSystem/x24/Favorite/Favorite.imageset/Favorite.pdf
@@ -1,87 +1,0 @@
-%PDF-1.7
-
-1 0 obj
-  << >>
-endobj
-
-2 0 obj
-  << /Length 3 0 R >>
-stream
-/DeviceRGB CS
-/DeviceRGB cs
-q
-1.000000 0.000000 -0.000000 1.000000 3.160980 3.743530 cm
-0.654199 0.654199 0.654199 scn
-7.823360 16.814558 m
-8.199115 17.717981 9.478912 17.717983 9.854667 16.814562 c
-11.490529 12.881474 l
-11.648937 12.500615 12.007109 12.240389 12.418278 12.207425 c
-16.664373 11.867018 l
-17.639692 11.788827 18.035173 10.571669 17.292084 9.935132 c
-14.057005 7.163942 l
-13.743737 6.895596 13.606929 6.474540 13.702637 6.073309 c
-14.691006 1.929840 l
-14.918032 0.978094 13.882655 0.225845 13.047646 0.735867 c
-9.412395 2.956264 l
-9.060376 3.171276 8.617653 3.171276 8.265635 2.956264 c
-4.630383 0.735867 l
-3.795373 0.225845 2.759995 0.978092 2.987021 1.929838 c
-3.975391 6.073309 l
-4.071099 6.474540 3.934291 6.895595 3.621023 7.163941 c
-0.385945 9.935131 l
--0.357145 10.571668 0.038332 11.788827 1.013651 11.867018 c
-5.259750 12.207425 l
-5.670919 12.240389 6.029090 12.500614 6.187498 12.881474 c
-7.823360 16.814558 l
-h
-f
-n
-Q
-
-endstream
-endobj
-
-3 0 obj
-  943
-endobj
-
-4 0 obj
-  << /Annots []
-     /Type /Page
-     /MediaBox [ 0.000000 0.000000 24.000000 24.000000 ]
-     /Resources 1 0 R
-     /Contents 2 0 R
-     /Parent 5 0 R
-  >>
-endobj
-
-5 0 obj
-  << /Kids [ 4 0 R ]
-     /Count 1
-     /Type /Pages
-  >>
-endobj
-
-6 0 obj
-  << /Pages 5 0 R
-     /Type /Catalog
-  >>
-endobj
-
-xref
-0 7
-0000000000 65535 f
-0000000010 00000 n
-0000000034 00000 n
-0000001033 00000 n
-0000001055 00000 n
-0000001228 00000 n
-0000001302 00000 n
-trailer
-<< /ID [ (some) (id) ]
-   /Root 6 0 R
-   /Size 7
->>
-startxref
-1361
-%%EOF
