--- Modules/Assets/Sources/Assets/Resources/Assets.xcassets/DesignSystem/x24/Favorite/Unfavorite.imageset/Contents.json
@@ -1,15 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "Unfavorite.pdf",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "properties" : {
-    "template-rendering-intent" : "template"
-  }
-}
