--- Modules/Assets/Sources/Assets/Resources/Assets.xcassets/DesignSystem/x24/Favorite/Unfavorite.imageset/Unfavorite.pdf
@@ -1,109 +1,0 @@
-%PDF-1.7
-
-1 0 obj
-  << >>
-endobj
-
-2 0 obj
-  << /Length 3 0 R >>
-stream
-/DeviceRGB CS
-/DeviceRGB cs
-q
-1.000000 0.000000 -0.000000 1.000000 2.807800 3.040710 cm
-0.579167 0.579167 0.579167 scn
-7.622566 17.487396 m
-8.203276 18.883591 10.181144 18.883598 10.761856 17.487400 c
-12.281930 13.832701 l
-12.382735 13.590336 12.610662 13.424737 12.872315 13.403761 c
-16.817867 13.087447 l
-18.325178 12.966607 18.936378 11.085545 17.787964 10.101806 c
-14.781868 7.526764 l
-14.582516 7.355998 14.495455 7.088054 14.556360 6.832725 c
-15.474771 2.982535 l
-15.825630 1.511658 14.225504 0.349091 12.935034 1.137306 c
-9.557090 3.200541 l
-9.333077 3.337367 9.051345 3.337367 8.827333 3.200541 c
-5.449391 1.137308 l
-4.158923 0.349094 2.558791 1.511650 2.909650 2.982533 c
-3.828061 6.832724 l
-3.888967 7.088053 3.801906 7.355998 3.602555 7.526764 c
-0.596460 10.101804 l
--0.551951 11.085541 0.059238 12.966606 1.566552 13.087447 c
-5.512107 13.403760 l
-5.773760 13.424737 6.001687 13.590336 6.102492 13.832701 c
-7.622566 17.487396 l
-h
-9.838535 17.103369 m
-9.599420 17.678270 8.785004 17.678272 8.545886 17.103367 c
-7.025813 13.448669 l
-6.781001 12.860069 6.227463 12.457901 5.592020 12.406959 c
-1.646465 12.090645 l
-1.025810 12.040888 0.774139 11.266332 1.247017 10.861261 c
-4.253111 8.286222 l
-4.737252 7.871504 4.948684 7.220781 4.800771 6.600698 c
-3.882360 2.750506 l
-3.737889 2.144852 4.396765 1.666149 4.928136 1.990709 c
-8.306079 4.053943 l
-8.850107 4.386233 9.534316 4.386232 10.078343 4.053943 c
-13.456287 1.990707 l
-13.987656 1.666151 14.646534 2.144852 14.502062 2.750508 c
-13.583651 6.600698 l
-13.435739 7.220782 13.647171 7.871504 14.131311 8.286222 c
-17.137407 10.861263 l
-17.610281 11.266331 17.358614 12.040888 16.737955 12.090645 c
-12.792402 12.406959 l
-12.156960 12.457901 11.603422 12.860069 11.358610 13.448669 c
-9.838535 17.103369 l
-h
-f*
-n
-Q
-
-endstream
-endobj
-
-3 0 obj
-  1768
-endobj
-
-4 0 obj
-  << /Annots []
-     /Type /Page
-     /MediaBox [ 0.000000 0.000000 24.000000 24.000000 ]
-     /Resources 1 0 R
-     /Contents 2 0 R
-     /Parent 5 0 R
-  >>
-endobj
-
-5 0 obj
-  << /Kids [ 4 0 R ]
-     /Count 1
-     /Type /Pages
-  >>
-endobj
-
-6 0 obj
-  << /Pages 5 0 R
-     /Type /Catalog
-  >>
-endobj
-
-xref
-0 7
-0000000000 65535 f
-0000000010 00000 n
-0000000034 00000 n
-0000001858 00000 n
-0000001881 00000 n
-0000002054 00000 n
-0000002128 00000 n
-trailer
-<< /ID [ (some) (id) ]
-   /Root 6 0 R
-   /Size 7
->>
-startxref
-2187
-%%EOF
