+++ Modules/Assets/Sources/Assets/Resources/Assets.xcassets/DesignSystem/x32/Favorite/Favorite.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "Variant=Favorite.pdf",
+      "filename" : "State=Inactive.svg",
       "idiom" : "universal"
     }
   ],
