--- Modules/Assets/Sources/Assets/Resources/Assets.xcassets/DesignSystem/x32/Favorite/Favorite.imageset/Variant=Favorite.pdf
@@ -1,109 +1,0 @@
-%PDF-1.7
-
-1 0 obj
-  << >>
-endobj
-
-2 0 obj
-  << /Length 3 0 R >>
-stream
-/DeviceRGB CS
-/DeviceRGB cs
-q
-1.000000 0.000000 -0.000000 1.000000 4.260132 3.751007 cm
-0.654199 0.654199 0.654199 scn
-9.651958 22.337450 m
-10.407739 24.219503 13.072054 24.219511 13.827836 22.337452 c
-15.705707 17.661156 l
-15.812989 17.394005 16.063721 17.211836 16.350948 17.192360 c
-21.378668 16.851454 l
-23.402163 16.714251 24.225471 14.180334 22.669083 12.879958 c
-18.801952 9.648941 l
-18.581028 9.464356 18.485256 9.169604 18.555492 8.890414 c
-19.784920 4.003431 l
-20.279724 2.036581 18.124252 0.470530 16.406565 1.548910 c
-12.138680 4.228327 l
-11.894859 4.381401 11.584936 4.381401 11.341116 4.228329 c
-7.073230 1.548910 l
-5.355547 0.470533 3.200068 2.036573 3.694873 4.003429 c
-4.924302 8.890412 l
-4.994538 9.169602 4.898766 9.464358 4.677842 9.648943 c
-0.810711 12.879960 l
--0.745684 14.180341 0.077642 16.714251 2.101130 16.851454 c
-7.128848 17.192360 l
-7.416075 17.211836 7.666807 17.394001 7.774088 17.661156 c
-9.651958 22.337450 l
-h
-12.435878 21.778481 m
-12.183951 22.405830 11.295845 22.405832 11.043918 21.778479 c
-9.166048 17.102184 l
-8.844205 16.300728 8.092009 15.754223 7.230322 15.695796 c
-2.202605 15.354891 l
-1.528105 15.309155 1.253669 14.464518 1.772464 14.031060 c
-5.639595 10.800042 l
-6.302372 10.246286 6.589684 9.362023 6.378975 8.524457 c
-5.149547 3.637474 l
-4.984612 2.981857 5.703104 2.459841 6.275668 2.819302 c
-10.543553 5.498718 l
-11.275015 5.957937 12.204782 5.957935 12.936243 5.498718 c
-17.204128 2.819300 l
-17.776688 2.459843 18.495182 2.981855 18.330246 3.637476 c
-17.100819 8.524461 l
-16.890110 9.362027 17.177422 10.246286 17.840199 10.800041 c
-21.707329 14.031058 l
-22.226130 14.464520 21.951687 15.309156 21.277193 15.354890 c
-16.249474 15.695796 l
-15.387787 15.754223 14.635591 16.300728 14.313748 17.102184 c
-12.435878 21.778481 l
-h
-f*
-n
-Q
-
-endstream
-endobj
-
-3 0 obj
-  1785
-endobj
-
-4 0 obj
-  << /Annots []
-     /Type /Page
-     /MediaBox [ 0.000000 0.000000 32.000000 32.000000 ]
-     /Resources 1 0 R
-     /Contents 2 0 R
-     /Parent 5 0 R
-  >>
-endobj
-
-5 0 obj
-  << /Kids [ 4 0 R ]
-     /Count 1
-     /Type /Pages
-  >>
-endobj
-
-6 0 obj
-  << /Pages 5 0 R
-     /Type /Catalog
-  >>
-endobj
-
-xref
-0 7
-0000000000 65535 f
-0000000010 00000 n
-0000000034 00000 n
-0000001875 00000 n
-0000001898 00000 n
-0000002071 00000 n
-0000002145 00000 n
-trailer
-<< /ID [ (some) (id) ]
-   /Root 6 0 R
-   /Size 7
->>
-startxref
-2204
-%%EOF
