+++ Modules/Assets/Sources/Assets/Resources/Assets.xcassets/DesignSystem/x32/Favorite/Unfavorite.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "Variant=Unfavorite.pdf",
+      "filename" : "State=Active.svg",
       "idiom" : "universal"
     }
   ],
