--- Modules/Assets/Sources/Assets/Resources/Assets.xcassets/DesignSystem/x32/Favorite/Unfavorite.imageset/Variant=Unfavorite.pdf
@@ -1,121 +1,0 @@
-%PDF-1.7
-
-1 0 obj
-  << >>
-endobj
-
-2 0 obj
-  << /Length 3 0 R >>
-stream
-/DeviceRGB CS
-/DeviceRGB cs
-q
-1.000000 0.000000 -0.000000 1.000000 4.260132 3.751007 cm
-0.654199 0.654199 0.654199 scn
-9.651958 22.337450 m
-10.407739 24.219503 13.072054 24.219511 13.827836 22.337452 c
-15.705707 17.661156 l
-15.741339 17.572428 15.792793 17.493073 15.856154 17.425940 c
-20.459538 22.029324 l
-20.752432 22.322216 21.227304 22.322216 21.520199 22.029324 c
-21.813091 21.736429 21.813091 21.261557 21.520199 20.968662 c
-2.020198 1.468662 l
-1.727305 1.175770 1.252431 1.175770 0.959538 1.468662 c
-0.666645 1.761557 0.666645 2.236429 0.959538 2.529324 c
-4.118815 5.688601 l
-4.924302 8.890412 l
-4.994538 9.169602 4.898766 9.464358 4.677842 9.648943 c
-0.810711 12.879960 l
--0.745684 14.180341 0.077642 16.714251 2.101130 16.851454 c
-7.128848 17.192360 l
-7.416075 17.211836 7.666807 17.394001 7.774088 17.661156 c
-9.651958 22.337450 l
-h
-6.185464 7.755250 m
-14.795327 16.365112 l
-14.591350 16.573009 14.426254 16.822020 14.313748 17.102184 c
-12.435878 21.778481 l
-12.183951 22.405830 11.295845 22.405832 11.043918 21.778479 c
-9.166048 17.102184 l
-8.844205 16.300728 8.092009 15.754223 7.230322 15.695796 c
-2.202605 15.354891 l
-1.528105 15.309155 1.253669 14.464518 1.772464 14.031060 c
-5.639595 10.800042 l
-6.302372 10.246286 6.589684 9.362023 6.378975 8.524457 c
-6.185464 7.755250 l
-h
-18.232843 15.561313 m
-19.640820 16.969290 l
-21.378668 16.851454 l
-23.402163 16.714251 24.225471 14.180334 22.669083 12.879958 c
-18.801952 9.648941 l
-18.581028 9.464356 18.485256 9.169604 18.555492 8.890414 c
-19.784920 4.003431 l
-20.279724 2.036581 18.124252 0.470530 16.406565 1.548910 c
-12.138680 4.228327 l
-11.894859 4.381401 11.584936 4.381401 11.341116 4.228329 c
-7.073230 1.548910 l
-6.155095 0.972496 5.111876 1.151623 4.415181 1.743650 c
-5.485904 2.814373 l
-5.711741 2.675592 6.009475 2.652184 6.275668 2.819302 c
-10.543553 5.498718 l
-11.275015 5.957937 12.204782 5.957935 12.936243 5.498718 c
-17.204128 2.819300 l
-17.776688 2.459843 18.495182 2.981855 18.330246 3.637476 c
-17.100819 8.524461 l
-16.890110 9.362027 17.177422 10.246286 17.840199 10.800041 c
-21.707329 14.031058 l
-22.226130 14.464520 21.951687 15.309156 21.277193 15.354890 c
-18.232843 15.561313 l
-h
-f*
-n
-Q
-
-endstream
-endobj
-
-3 0 obj
-  2169
-endobj
-
-4 0 obj
-  << /Annots []
-     /Type /Page
-     /MediaBox [ 0.000000 0.000000 32.000000 32.000000 ]
-     /Resources 1 0 R
-     /Contents 2 0 R
-     /Parent 5 0 R
-  >>
-endobj
-
-5 0 obj
-  << /Kids [ 4 0 R ]
-     /Count 1
-     /Type /Pages
-  >>
-endobj
-
-6 0 obj
-  << /Pages 5 0 R
-     /Type /Catalog
-  >>
-endobj
-
-xref
-0 7
-0000000000 65535 f
-0000000010 00000 n
-0000000034 00000 n
-0000002259 00000 n
-0000002282 00000 n
-0000002455 00000 n
-0000002529 00000 n
-trailer
-<< /ID [ (some) (id) ]
-   /Root 6 0 R
-   /Size 7
->>
-startxref
-2588
-%%EOF
